requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pypdf>=4.2.0
python-docx>=1.1.2
openai>=1.40.0
python-dotenv>=1.0.1
